backend/requirements-backend.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
numpy
Pillow
